ipsec: move pluto to host
this is done to allow co-existance with host configured ipsec tunnels

To avoid mounting the whole host fs, only the configuration files are
shared. including NSS DB and the whole ipsec configuration

hostPID is used, so that the pid file make sense from the host as well.
that required moving the termination logic to a preStop lifecycle step.

the rest stays pretty much the same :-)